+++ docs/generate.js
@@ -27,6 +27,7 @@ function generate(pathname) {
 		if (pathname.match(/\.md$/)) {
 			var outputFilename = pathname.replace(/\.md$/, ".html")
 			var markdown = fs.readFileSync(pathname, "utf-8")
+			var anchors = {}
 			var fixed = markdown
 				.replace(/`((?:\S| -> |, )+)(\|)(\S+)`/gim, function(match, a, b, c) { // fix pipes in code tags
 					return "<code>" + (a + b + c).replace(/\|/g, "&#124;") + "</code>"
@@ -48,10 +49,16 @@ function generate(pathname) {
 			var title = fixed.match(/^#([^\n\r]+)/i) || []
 			var html = layout
 				.replace(/<title>Mithril\.js<\/title>/, "<title>" + title[1] + " - Mithril.js</title>")
-				.replace(/\[version\]/, version) // update version
+				.replace(/\[version\]/g, version) // update version
 				.replace(/\[body\]/, markedHtml)
 				.replace(/<h(.) id="([^"]+?)">(.+?)<\/h.>/gim, function(match, n, id, text) { // fix anchors
 					var anchor = text.toLowerCase().replace(/<(\/?)code>/g, "").replace(/<a.*?>.+?<\/a>/g, "").replace(/\.|\[|\]|&quot;|\/|\(|\)/g, "").replace(/\s/g, "-");
+
+					if(anchor in anchors) {
+						anchor += ++anchors[anchor]
+					} else {
+						anchors[anchor] = 0;
+					}
 
 					return `<h${n} id="${anchor}"><a href="#${anchor}">${text}</a></h${n}>`;
 				})